backend/requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
pydantic==2.12.4
yfinance==0.2.66
apscheduler==3.10.4
pytz==2024.1
alpha_vantage==2.3.1
python-dotenv==1.0.1
sqlalchemy==2.0.44
PyYAML==6.0.3